solve ES sorting issues
... ES is quite different from other datasources given that its auto-generated ID does not sit within the document, this caused some issues when the query generator attempted to request that search results be auto-sorted on `id` by default ... this is good behavior to emulate for loopback but there is no `id` field present in the docs for ES!

It worked fine when explicitly populating an ID field like `seq` or `id` with values inside the doc but that use-case is not for everyone, most ES users depend on auto-generation for the ID and even when they specify it explicitly, they save it in the `_id` field separate from the doc ... folks don't think of also saving it inside the doc ... and they shouldn't have to just because of this connector.

Now we hopefully have a decent solution to handle all situations.